rust-server-cleanup(mm-module,non_mechanical_provable): extract multimodal integration
