+++ python/sglang/srt/rust_server/__init__.py
@@ -1,13 +1,13 @@
 """Embedded Rust server integration."""
 
-from sglang.srt.rust_server.server import (
+from sglang.srt.rust_server.multimodal import (
     NATIVE_MM_FAMILIES,
     NativeMmFamily,
     NativeMmHost,
     NativeMmSpec,
-    RustServer,
     native_mm_family_for,
 )
+from sglang.srt.rust_server.server import RustServer
 
 __all__ = [
     "NATIVE_MM_FAMILIES",
